fix: CategoryFilter is presentational
