+++ src/containers/BooksList.js
@@ -2,37 +2,42 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import Book from '../components/Book';
-import { removeBook } from '../actions';
-import CategoryFilter from './CategoryFilter';
+import { removeBook, changeFilter } from '../actions';
+import CategoryFilter from '../components/CategoryFilter';
 
 const mapStateToProps = state => {
   const { books, filter } = state;
+
   return { books, filter };
 };
 
-const booksCollection = books => {
-  const result = [];
-  Object.keys(books).forEach(k => {
-    result.push({ id: k, ...books[k] });
-  });
-  return result;
-};
+const booksCollection = books => (
+  Object.keys(books).map(k => ({ id: k, ...books[k] })
+);
 
 const mapDispatchToProps = dispatch => ({
   removeBook: id => {
     dispatch(removeBook(id));
   },
+  changeFilter: filter => {
+    dispatch(changeFilter(filter));
+  },
 });
 
-const Component = ({ books, filter, removeBook }) => {
+const Component = ({ books, filter, removeBook, changeFilter }) => {
+  const handleFilterChange = e => changeFilter(e.target.value);
+
   const filteredBooks = () => {
     const booksInArray = booksCollection(books);
+
     if (filter === 'all') return booksInArray;
+
     return booksInArray.filter(book => book.category === filter);
   };
+
   return (
     <>
-      <CategoryFilter />
+      <CategoryFilter handleFilterChange={handleFilterChange} />
       <table id="books-table" className="panel">
         <thead>
           <tr>
@@ -56,6 +61,7 @@ Component.propTypes = {
   books: PropTypes.objectOf(PropTypes.object).isRequired,
   removeBook: PropTypes.func.isRequired,
   filter: PropTypes.string.isRequired,
+  changeFilter: PropTypes.func.isRequired,
 };
 
 const BooksList = connect(mapStateToProps, mapDispatchToProps)(Component);
